Develop

+++ src/Controller/TaskController.php
@@ -4,24 +4,51 @@ namespace App\Controller;
 
 use App\Entity\Task;
 use App\Form\TaskType;
-use Symfony\Component\Routing\Annotation\Route;
+use App\Repository\TaskRepository;
+use Doctrine\ORM\Mapping\Id;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Security;
 
 class TaskController extends AbstractController
 {
     /**
+     * @return [type]
+     *
      * @Route("/tasks", name="task_list")
      */
-    public function listAction()
+    public function listTask(TaskRepository $taskRepository)
     {
-        return $this->render('task/list.html.twig', ['tasks' => $this->getDoctrine()->getRepository('App:Task')->findAll()]);
+        $tasks = $taskRepository->findAll();
+
+        return $this->render(
+            'task/list.html.twig',
+            ['tasks' => $tasks]
+        );
     }
 
     /**
+     * @return [type]
+     *
+     * @Route("/tasks/done", name="task_list_done", methods={"GET"})
+     */
+    public function tasksDone(TaskRepository $taskRepository)
+    {
+        $tasks = $taskRepository->findBy(['isDone' => 'True']);
+
+        return $this->render(
+            'task/list_done.html.twig',
+            ['tasks' => $tasks]
+        );
+    }
+
+    /**
+     * @return [type]
+     *
      * @Route("/tasks/create", name="task_create")
      */
-    public function createAction(Request $request)
+    public function createTask(Request $request, Security $security)
     {
         $task = new Task();
         $form = $this->createForm(TaskType::class, $task);
@@ -29,11 +56,10 @@ class TaskController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $task->setUser($security->getUser());
             $em = $this->getDoctrine()->getManager();
-
             $em->persist($task);
             $em->flush();
-
             $this->addFlash('success', 'La tâche a été bien été ajoutée.');
 
             return $this->redirectToRoute('task_list');
@@ -43,15 +69,26 @@ class TaskController extends AbstractController
     }
 
     /**
+     * @param mixed $id
+     *
+     * @return [type]
+     *
      * @Route("/tasks/{id}/edit", name="task_edit")
      */
-    public function editAction(Task $task, Request $request)
+    public function editTask($id, Request $request, TaskRepository $taskRepository)
     {
+        $task = $taskRepository->find($id);
+        if (!($task)) {
+            $this->addFlash('error', "Cette tâche n'existe pas");
+
+            return $this->RedirectToRoute('task_list');
+        }
+
         $form = $this->createForm(TaskType::class, $task);
 
         $form->handleRequest($request);
 
-        if ($form->isSubmitted() &&  $form->isValid()) {
+        if ($form->isSubmitted() && $form->isValid()) {
             $this->getDoctrine()->getManager()->flush();
 
             $this->addFlash('success', 'La tâche a bien été modifiée.');
@@ -66,27 +103,42 @@ class TaskController extends AbstractController
     }
 
     /**
+     * @return [type]
+     *
      * @Route("/tasks/{id}/toggle", name="task_toggle")
      */
-    public function toggleTaskAction(Task $task)
+    public function toggleTask(Task $task)
     {
         $task->toggle(!$task->isDone());
         $this->getDoctrine()->getManager()->flush();
-
-        $this->addFlash('success', sprintf('La tâche %s a bien été marquée comme faite.', $task->getTitle()));
+        if ($task->isDone()) {
+            $this->addFlash('success', sprintf('La tâche %s a bien été marquée comme faite.', $task->getTitle()));
+        } else {
+            $this->addFlash('success', sprintf('La tâche %s reste à faire.', $task->getTitle()));
+        }
 
         return $this->redirectToRoute('task_list');
     }
 
     /**
+     * @param mixed $id
+     *
+     * @return [type]
+     *
      * @Route("/tasks/{id}/delete", name="task_delete")
      */
-    public function deleteTaskAction(Task $task)
+    public function deleteTask($id, TaskRepository $taskRepository)
     {
+        $task = $taskRepository->find($id);
+        if (!($task)) {
+            $this->addFlash('error', "Cette tâche n'existe pas");
+
+            return $this->RedirectToRoute('task_list');
+        }
+        $this->denyAccessUnlessGranted('TASK_DELETE', $task, "Vous n'etes pas l'auteur de cette tâche");
         $em = $this->getDoctrine()->getManager();
         $em->remove($task);
         $em->flush();
-
         $this->addFlash('success', 'La tâche a bien été supprimée.');
 
         return $this->redirectToRoute('task_list');
